+++ src/pages/about/index.js
@@ -1,6 +1,7 @@
-import Taro, { Component } from '@tarojs/taro';
-import { View } from '@tarojs/components';
-import './index.scss';
+import Taro, { Component } from '@tarojs/taro'
+import { View } from '@tarojs/components'
+import SearchHeader from '../../components/SearchHeader'
+import './index.scss'
 
 export default class About extends Component {
   config = {
@@ -13,11 +14,12 @@ export default class About extends Component {
 
   render() {
     return (
-      <View className='about-page'>
-        <View>云南云农科技有限公司</View>
+      <View>
+        <SearchHeader></SearchHeader>
+        {/* <View>云南云农科技有限公司</View>
         <View>公司介绍1</View>
         <View>公司介绍2</View>
-        <View>公司介绍3</View>
+        <View>公司介绍3</View> */}
       </View>
     )
   }